+++ exts/meta/_utils.py
@@ -1,10 +1,13 @@
+from core import db
+
+
 async def getDisabledCommands(bot, guildId):
     if bot.cache.disabled.get(guildId) is None:
-        dbDisabled = await bot.db.fetch_all(
-            "SELECT command FROM disabled WHERE guildId=:id", values={"id": guildId}
-        )
+        dbDisabled = await db.Disabled.filter(guild_id=guildId)
+
         try:
-            bot.cache.disabled.extend(guildId, [c[0] for c in dbDisabled])
+            bot.cache.disabled.extend(guildId, [c.command for c in dbDisabled])
         except ValueError:
             return []
+
     return bot.cache.disabled.get(guildId, [])